Enable ETW error logging for S.SM.Federation (#4391)
