+++ src/Common/src/Internals/System/Runtime/AssertHelper.cs
@@ -1,0 +1,17 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics;
+
+namespace System.Runtime
+{
+    internal static class AssertHelper
+    {
+        internal static void FireAssert(string message)
+        {
+            Debug.Assert(false, message);
+        }
+    }
+}
+
