+++ src/Common/src/Internals/System/Runtime/DiagnosticStrings.cs
@@ -1,0 +1,19 @@
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime
+{
+    internal static partial class DiagnosticStrings
+    {
+        internal const string DataTag = "Data";
+        internal const string DataItemsTag = "DataItems";
+        internal const string ExceptionStringTag = "ExceptionString";
+        internal const string ExceptionTag = "Exception";
+        internal const string ExceptionTypeTag = "ExceptionType";
+        internal const string KeyTag = "Key";
+        internal const string InnerExceptionTag = "InnerException";
+        internal const string MessageTag = "Message";
+        internal const string StackTraceTag = "StackTrace";
+        internal const string ValueTag = "Value";
+    }
+}
