+++ src/Common/src/Internals/System/Runtime/Fx.cs
@@ -1,0 +1,27 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics;
+
+namespace System.Runtime
+{
+    internal static partial class Fx
+    {
+        [Conditional("DEBUG")]
+        public static void Assert(bool condition, string description)
+        {
+            if (!condition)
+            {
+                Assert(description);
+            }
+        }
+
+        [Conditional("DEBUG")]
+        public static void Assert(string description)
+        {
+            AssertHelper.FireAssert(description);
+        }
+
+    }
+}
