+++ src/Common/src/System/ServiceModel/AddressingStrings.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.ServiceModel
+{
+    internal static partial class AddressingStrings
+    {
+        // String constants
+        public const string EndpointUnavailable = "EndpointUnavailable";
+        public const string ActionNotSupported = "ActionNotSupported";
+        public const string EndpointReferenceType = "EndpointReferenceType";
+        public const string Request = "Request";
+        public const string DestinationUnreachable = "DestinationUnreachable";
+        public const string AnonymousUri = "http://schemas.microsoft.com/2005/12/ServiceModel/Addressing/Anonymous";
+        public const string NoneUri = "http://schemas.microsoft.com/2005/12/ServiceModel/Addressing/None";
+        public const string IndigoNamespace = "http://schemas.microsoft.com/serviceModel/2004/05/addressing";
+        public const string ChannelTerminated = "ChannelTerminated";
+    }
+}
