+++ src/Common/src/System/ServiceModel/Diagnostics/ExceptionUtility.cs
@@ -1,0 +1,23 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics.Tracing;
+
+namespace System.ServiceModel.Diagnostics
+{
+    internal partial class ExceptionUtility
+    {
+        public Exception ThrowHelperError(Exception exception)
+        {
+            return ThrowHelper(exception, EventLevel.Error);
+        }
+
+        internal Exception ThrowHelper(Exception exception, EventLevel eventLevel)
+        {
+            FxTrace.Exception.TraceEtwException(exception, eventLevel);
+            return exception;
+        }
+
+    }
+}
