+++ src/System.Private.ServiceModel/src/Internals/System/Runtime/DiagnosticStrings.cs
@@ -2,33 +2,22 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
 namespace System.Runtime
 {
-    internal static class DiagnosticStrings
+    internal static partial class DiagnosticStrings
     {
         internal const string AppDomain = "AppDomain";
         internal const string ChannelTag = "Channel";
         internal const string Description = "Description";
-        internal const string DataTag = "Data";
-        internal const string DataItemsTag = "DataItems";
         internal const string DescriptionTag = "Description";
-        internal const string ExceptionTag = "Exception";
-        internal const string ExceptionTypeTag = "ExceptionType";
-        internal const string ExceptionStringTag = "ExceptionString";
         internal const string ExtendedDataTag = "ExtendedData";
-        internal const string InnerExceptionTag = "InnerException";
-        internal const string KeyTag = "Key";
-        internal const string MessageTag = "Message";
         internal const string NamespaceTag = "xmlns";
         internal const string NativeErrorCodeTag = "NativeErrorCode";
         internal const string Separator = ":";
         internal const string SeverityTag = "Severity";
         internal const string SourceTag = "Source";
-        internal const string StackTraceTag = "StackTrace";
         internal const string Task = "Task";
         internal const string TraceCodeTag = "TraceIdentifier";
         internal const string TraceRecordTag = "TraceRecord";
-        internal const string ValueTag = "Value";
     }
 }
